chore: un-nest CLIs, APIs, permissions and DB layer
