+++ master/.mockery.yaml
@@ -70,6 +70,12 @@ packages:
           mockname: AllocationService
           dir: internal/mocks/allocationmocks
           outpkg: allocationmocks
+  github.com/determined-ai/determined/master/internal/token:
+    interfaces:
+      TokenAuthZ:
+        config:
+          filename: token_authz_iface.go
+          mockname: TokenAuthZ
   k8s.io/client-go/kubernetes/typed/core/v1:
     interfaces:
       PodInterface:
