+++ master/internal/user/authz_iface.go
@@ -3,8 +3,6 @@ package user
 import (
 	"context"
 
-	"github.com/uptrace/bun"
-
 	"github.com/determined-ai/determined/master/internal/authz"
 	"github.com/determined-ai/determined/master/pkg/model"
 )
@@ -66,14 +64,6 @@ type UserAuthZ interface {
 	) error
 	// POST /api/v1/users/setting/reset
 	CanResetUsersOwnSettings(ctx context.Context, curUser model.User) error
-
-	// POST /api/v1/users/:user_id/token
-	CanCreateAccessToken(ctx context.Context, curUser, targetUser model.User) error
-	// GET /api/v1/user/tokens
-	CanGetAccessTokens(ctx context.Context, curUser model.User, query *bun.SelectQuery,
-		targetUserID model.UserID) (*bun.SelectQuery, error)
-	// PATCH /api/v1/users/token/:token_id
-	CanUpdateAccessToken(ctx context.Context, curUser model.User, targetTokenUserID model.UserID) error
 }
 
 // AuthZProvider is the authz registry for `user` package.
