+++ master/internal/user/authz_permissive.go
@@ -3,8 +3,6 @@ package user
 import (
 	"context"
 
-	"github.com/uptrace/bun"
-
 	"github.com/determined-ai/determined/master/pkg/model"
 )
 
@@ -121,32 +119,6 @@ func (p *UserAuthZPermissive) CanResetUsersOwnSettings(
 ) error {
 	_ = (&UserAuthZRBAC{}).CanResetUsersOwnSettings(ctx, curUser)
 	return (&UserAuthZBasic{}).CanResetUsersOwnSettings(ctx, curUser)
-}
-
-// CanCreateAccessToken calls RBAC authz but enforces basic authz.
-func (p *UserAuthZPermissive) CanCreateAccessToken(
-	ctx context.Context, curUser, targetUser model.User,
-) error {
-	_ = (&UserAuthZRBAC{}).CanCreateAccessToken(ctx, curUser, targetUser)
-	return (&UserAuthZBasic{}).CanCreateAccessToken(ctx, curUser, targetUser)
-}
-
-// CanGetAccessTokens calls RBAC authz but enforces basic authz.
-func (p *UserAuthZPermissive) CanGetAccessTokens(
-	ctx context.Context, curUser model.User, query *bun.SelectQuery, targetUserID model.UserID,
-) (*bun.SelectQuery, error) {
-	_, _ = (&UserAuthZRBAC{}).CanGetAccessTokens(ctx, curUser, query, targetUserID)
-	return (&UserAuthZBasic{}).CanGetAccessTokens(ctx, curUser, query, targetUserID)
-}
-
-// CanUpdateAccessToken calls RBAC authz but enforces basic authz.
-func (p *UserAuthZPermissive) CanUpdateAccessToken(
-	ctx context.Context,
-	curUser model.User,
-	targetTokenUserID model.UserID,
-) error {
-	_ = (&UserAuthZRBAC{}).CanUpdateAccessToken(ctx, curUser, targetTokenUserID)
-	return (&UserAuthZBasic{}).CanUpdateAccessToken(ctx, curUser, targetTokenUserID)
 }
 
 func init() {
